refactoring and fixed bug

+++ CPP/Header/CommandTable.h
@@ -2,29 +2,27 @@
 #define COMMAND_TABLE_HEADER
 #include <stdio.h>
 #include <unordered_map>
+#include <climits>
 using namespace std;
 
 void EndProgram(bool& running);
 
-void MOVValueIndexLeft(int& valueIndex);
+void MOVValueIndexLeft(unsigned int& valueIndex);
 
-void MOVValueIndexRight(int& valueIndex);
+void MOVValueIndexRight(unsigned int& valueIndex);
 
-void INCRValue(long long valueLine[], int valueIndex);
+void INCRValue(int valueLine[], unsigned int valueIndex);
 
-void DECRValue(long long valueLine[], int valueIndex);
+void DECRValue(int valueLine[], unsigned int valueIndex);
 
-void PrintValueByASCII(long long valueLine[], int valueIndex);
+void PrintValueByASCII(int valueLine[], unsigned int valueIndex);
 
-void InputValueByASCII(long long valueLine[], int valueIndex);
+void InputValueByASCII(int valueLine[], unsigned int valueIndex);
 
-void JMPCommandIndexToTail(unordered_map<int,int> loopMap, int& commandIndex);
+void JMPCommandIndexToTail(unordered_map<int,int> loopMap, unsigned int& commandIndex);
 
-void JMPCommandIndexToHead(unordered_map<int , int> loopMap, int& commandIndex);
+void JMPCommandIndexToHead(unordered_map<int , int> loopMap, unsigned int& commandIndex);
 
-
-
-void PrintValueByInteger(long long valueLine[], int valueIndex);
-
+void PrintValueByInteger(int valueLine[], unsigned int valueIndex);
 
 #endif // !COMMAND_TABLE_HEADER
